+++ packages/gdl-frontend/components/BookListSection/coverWidths.js
